@@ -5,35 +5,37 @@
   "homepage": "./ui",
   "packageManager": "npm@9.8.1",
   "dependencies": {
-    "@date-io/moment": "^2.8.0",
-    "@material-ui/core": "^4.9.5",
-    "@material-ui/pickers": "^3.2.10",
-    "@popperjs/core": "^2.0.0",
+    "@date-io/date-fns": "^2.17.0",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.15.11",
+    "@mui/x-date-pickers": "^6.19.5",
+    "@popperjs/core": "^2.11.8",
     "@trendmicro/react-sidenav": "^0.5.0",
-    "ace-builds": "^1.4.8",
-    "axios": "^0.21.1",
-    "axios-progress-bar": "^1.2.0",
-    "bootstrap": "^4.6.1",
+    "ace-builds": "^1.32.6",
+    "axios": "^1.6.7",
+    "bootstrap": "^5.3.3",
     "buffer": "^6.0.3",
     "convert-units": "^2.3.4",
-    "css-loader": "^3.4.1",
+    "css-loader": "^6.10.0",
+    "date-fns": "^2.17.0",
     "event-source-polyfill": "^1.0.31",
     "font-awesome": "^4.7.0",
+    "history": "^5.3.0",
     "joi-validation": "^2.0.0",
     "lodash": "^4.17.21",
-    "lossless-json": "^2.0.1",
-    "moment": "^2.24.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-ace": "^9.1.2",
-    "react-bootstrap": "^1.3.0",
-    "react-datepicker": "^3.1.3",
-    "react-dom": "^16.13.1",
-    "react-helmet": "^6.0.0",
-    "react-router-dom": "^5.1.2",
+    "lossless-json": "^4.0.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-ace": "^10.1.0",
+    "react-bootstrap": "^2.10.1",
+    "react-datepicker": "^6.1.0",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
+    "react-router-dom": "^6.22.1",
     "react-scripts": "^5.0.1",
-    "react-timeago": "^6.2.1",
-    "react-toastify": "^6.0.6"
+    "react-timeago": "^7.2.0",
+    "react-toastify": "^10.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -55,23 +57,21 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@babel/core": "^7.20.2",
-    "@babel/eslint-parser": "^7.19.1",
-    "@testing-library/jest-dom": "^5.11.2",
-    "@testing-library/react": "^10.4.7",
-    "@testing-library/user-event": "^12.1.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "eslint": "^8.27.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.2.1",
+    "@babel/core": "^7.23.9",
+    "@babel/eslint-parser": "^7.23.10",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "jest-dom": "^4.0.0",
     "joi-browser": "^13.4.0",
-    "prettier": "^2.7.1",
-    "react-test-renderer": "^16.13.1",
-    "sass": "^1.56.0",
-    "style-loader": "^1.1.2",
-    "webpack": "4.42.0"
+    "prettier": "^3.2.5",
+    "react-test-renderer": "^18.2.0",
+    "sass": "^1.71.1",
+    "style-loader": "^3.3.4",
+    "webpack": "5.90.3"
   },
   "proxy": "http://akhq:8080"
 }
