@@ -1,13 +1,15 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.scss';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import prefix from './prefix';
+import { createRoot } from 'react-dom/client';
 
 let pathPrefix = prefix() + '/ui';
 
-ReactDOM.render(<App pathPrefix={pathPrefix} />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<App pathPrefix={pathPrefix} />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
