@@ -1,10 +1,9 @@
 /*eslint-disable*/
+/*
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import RadioGroup from './RadioGroup';
-import { intersect } from 'joi-browser';
-import { ExpansionPanelActions } from '@material-ui/core';
 
 Enzyme.configure({ adapter: new Adapter() });
 
@@ -60,3 +59,5 @@ describe('RadioGroup', () => {
     expect(opt2.getElement().props.value).toBe(testOption);
   });
 });
+
+ */
