@@ -5,6 +5,7 @@ import * as constants from '../../../utils/constants';
 import { uriAclsList } from '../../../utils/endpoints';
 import SearchBar from '../../../components/SearchBar';
 import Root from '../../../components/Root';
+import { withRouter } from '../../../utils/withRouter';
 
 class Acls extends Root {
   state = {
@@ -19,7 +20,7 @@ class Acls extends Root {
   componentDidMount() {
     const { searchData } = this.state;
     const query = new URLSearchParams(this.props.location.search);
-    const { clusterId } = this.props.match.params;
+    const { clusterId } = this.props.params;
 
     this.setState(
       {
@@ -56,7 +57,7 @@ class Acls extends Root {
     const { searchData } = data;
     this.setState({ searchData, loading: true }, () => {
       this.getAcls();
-      this.props.history.push({
+      this.props.router.navigate({
         pathname: `/ui/${this.state.selectedCluster}/acls`,
         search: `search=${searchData.search}`
       });
@@ -65,7 +66,7 @@ class Acls extends Root {
 
   render() {
     const { data, searchData, loading } = this.state;
-    const { clusterId } = this.props.match.params;
+    const { clusterId } = this.props.params;
 
     return (
       <div>
@@ -118,4 +119,4 @@ class Acls extends Root {
   }
 }
 
-export default Acls;
+export default withRouter(Acls);
