@@ -1,6 +1,5 @@
 import React from 'react';
 import { uriKsqlDBExecuteStatement } from '../../../utils/endpoints';
-import { withRouter } from 'react-router-dom';
 import Header from '../../Header/Header';
 import 'ace-builds/webpack-resolver';
 import 'ace-builds/src-noconflict/mode-json';
@@ -9,6 +8,7 @@ import { toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import Root from '../../../components/Root';
 import AceEditor from 'react-ace';
+import { withRouter } from '../../../utils/withRouter';
 
 class KsqlDBStatement extends Root {
   state = {
