@@ -4,6 +4,7 @@ import Table from '../../../components/Table';
 import * as constants from '../../../utils/constants';
 import { uriNodes, uriNodePartitions } from '../../../utils/endpoints';
 import Root from '../../../components/Root';
+import { withRouter } from '../../../utils/withRouter';
 
 class NodesList extends Root {
   state = {
@@ -18,14 +19,14 @@ class NodesList extends Root {
 
   async getNodes() {
     let nodes = [];
-    const { clusterId } = this.props.match.params;
+    const { clusterId } = this.props.params;
     nodes = await this.getApi(uriNodes(clusterId));
     this.handleData(nodes.data);
     this.setState({ selectedCluster: clusterId });
   }
 
   handleData(nodes) {
-    const { clusterId } = this.props.match.params;
+    const { clusterId } = this.props.params;
     let tableNodes = {};
     const setState = () => {
       this.setState({ data: Object.values(tableNodes), loading: false });
@@ -121,4 +122,4 @@ class NodesList extends Root {
   }
 }
 
-export default NodesList;
+export default withRouter(NodesList);
