@@ -22,6 +22,7 @@ import DateTime from '../../../components/DateTime';
 import { getClusterUIOptions } from '../../../utils/functions';
 import { handlePageChange, getPageNumber } from './../../../utils/pagination';
 import PageSize from '../../../components/PageSize';
+import { withRouter } from '../../../utils/withRouter';
 
 class TopicList extends Root {
   state = {
@@ -56,7 +57,7 @@ class TopicList extends Root {
   componentDidMount() {
     this._initializeVars(() => {
       this.getTopics();
-      this.props.history.replace({
+      this.props.router.navigate({
         pathname: `/ui/${this.state.selectedCluster}/topic`,
         search: this.props.location.search
       });
@@ -75,7 +76,7 @@ class TopicList extends Root {
   }
 
   async _initializeVars(callBackFunction) {
-    const { clusterId } = this.props.match.params;
+    const { clusterId } = this.props.params;
     const query = new URLSearchParams(this.props.location.search);
     const { searchData, keepSearch } = this.state;
     let { pageNumber } = this.state;
@@ -94,8 +95,8 @@ class TopicList extends Root {
         topicListView: query.get('topicListView')
           ? query.get('topicListView')
           : uiOptions && uiOptions.topic && uiOptions.topic.defaultView
-          ? uiOptions.topic.defaultView
-          : searchData.topicListView
+            ? uiOptions.topic.defaultView
+            : searchData.topicListView
       };
       pageNumber = query.get('page') ? parseInt(query.get('page')) : parseInt(pageNumber);
       currentPageSize = query.get('uiPageSize')
@@ -152,7 +153,7 @@ class TopicList extends Root {
       const { topicListView } = this.state.searchData;
       this.getTopics();
       this.handleKeepSearchChange(data.keepSearch);
-      this.props.history.push({
+      this.props.router.navigate({
         pathname: `/ui/${this.state.selectedCluster}/topic`,
         search: `search=${searchData.search}&topicListView=${topicListView}&page=${this.state.pageNumber}`
       });
@@ -165,7 +166,7 @@ class TopicList extends Root {
     this.setState({ pageNumber: pageNumber }, () => {
       const { search, topicListView } = this.state.searchData;
       this.getTopics();
-      this.props.history.push({
+      this.props.router.navigate({
         pathname: `/ui/${this.state.selectedCluster}/topic`,
         search: `search=${search}&topicListView=${topicListView}&page=${pageNumber}`
       });
@@ -177,7 +178,7 @@ class TopicList extends Root {
     this.setState({ currentPageSize: value, pageNumber: pageNumber }, () => {
       const { search, topicListView } = this.state.searchData;
       this.getTopics();
-      this.props.history.push({
+      this.props.router.navigate({
         pathname: `/ui/${this.state.selectedCluster}/topic`,
         search: `search=${search}&topicListView=${topicListView}&uiPageSize=${value}`
       });
@@ -345,7 +346,7 @@ class TopicList extends Root {
         ? uiOptions.topicData.dateTimeFormat
         : constants.SETTINGS_VALUES.TOPIC_DATA.DATE_TIME_FORMAT.RELATIVE;
     const roles = this.state.roles || {};
-    const { clusterId } = this.props.match.params;
+    const { clusterId } = this.props.params;
 
     const topicCols = [
       {
@@ -566,4 +567,4 @@ class TopicList extends Root {
   }
 }
 
-export default TopicList;
+export default withRouter(TopicList);
